refactor backward compatibility utils
