+++ .c8rc
@@ -5,7 +5,11 @@
   "functions": 100,
   "statements": 100,
   "exclude": ["/node_modules/", "/src/prettier-comments/"],
-  "include": ["src/**/*.js", "!src/prettier-comments/**/*.js"],
+  "include": [
+    "src/**/*.js",
+    "!src/prettier-comments/**/*.js",
+    "!src/common/backward-compatibility.js"
+  ],
   "reporter": ["lcov", "text"],
   "temp-dir": "./coverage/"
 }
