+++ src/comments/handlers/handleModifierInvocationComments.js
@@ -1,5 +1,5 @@
 import { util } from 'prettier';
-import { getNextNonSpaceNonCommentCharacter } from '../../common/util.js';
+import { getNextNonSpaceNonCommentCharacter } from '../../common/backward-compatibility.js';
 
 const { addLeadingComment, addTrailingComment, addDanglingComment } = util;
