+++ src/common/printer-helpers.js
@@ -1,12 +1,12 @@
 import { doc } from 'prettier';
+import { isPrettier2 } from './util.js';
 import {
   isFirst,
   isLast,
   isNextLineEmpty,
-  isPrettier2,
   next,
   previous
-} from './util.js';
+} from './backward-compatibility.js';
 
 const { group, indent, join, line, softline, hardline } = doc.builders;
