+++ src/prettier-comments/language-js/comments.js
@@ -1,5 +1,5 @@
 import { util } from "prettier";
-import { getNextNonSpaceNonCommentCharacter } from "../../common/util.js";
+import { getNextNonSpaceNonCommentCharacter } from "../../common/backward-compatibility.js";
 
 const {
   addLeadingComment,
